+++ controllers/operator/construct/backup_construction.go
@@ -28,8 +28,8 @@ const (
 	healthEndpointPortEnv             = "HEALTH_ENDPOINT_PORT"
 	backupDaemonReadinessProbeCommand = "/opt/scripts/backup-daemon-readiness-probe"
 	backupDaemonLivenessProbeCommand  = "/opt/scripts/backup-daemon-liveness-probe.sh"
-	// mmsHome corresponds to MMS_HOME in the Ops Manager Dockerfile.
-	mmsHome = "/mongodb-ops-manager"
+	// MMSHome corresponds to MMS_HOME in the Ops Manager Dockerfile.
+	MMSHome = "/mongodb-ops-manager"
 )
 
 // BackupDaemonStatefulSet fully constructs the Backup StatefulSet.
@@ -107,7 +107,7 @@ func backupDaemonStatefulSetFunc(opts OpsManagerStatefulSetOptions) statefulset.
 		caCertVolume := statefulset.CreateVolumeFromConfigMap("ops-manager-ca", opts.OpsManagerCaName)
 		caVolumeFunc = podtemplatespec.WithVolume(caCertVolume)
 		caVolumeMountFunc = container.WithVolumeMounts([]corev1.VolumeMount{{
-			MountPath: fmt.Sprintf("%s/%s", mmsHome, caCertVolume.Name),
+			MountPath: fmt.Sprintf("%s/%s", MMSHome, caCertVolume.Name),
 			Name:      caCertVolume.Name,
 			ReadOnly:  true,
 		}})
@@ -115,7 +115,7 @@ func backupDaemonStatefulSetFunc(opts OpsManagerStatefulSetOptions) statefulset.
 		// It will add each X.509 public key certificate into JVM's trust store
 		// with unique "mongodb_operator_added_trust_ca_$RANDOM" alias
 		// See: https://jira.mongodb.org/browse/HELP-25872 for more details.
-		postStartScript := fmt.Sprintf(`awk -v cmd="%s/jdk/bin/keytool -noprompt -storepass changeit -import -trustcacerts -alias mongodb_operator_added_trust_ca_${RANDOM} -keystore %s/jdk/lib/security/cacerts" '/BEGIN/{close(cmd)};{print | cmd}' 2>&1 < %s/%s/ca-pem`, mmsHome, mmsHome, mmsHome, caCertVolume.Name)
+		postStartScript := fmt.Sprintf(`awk -v cmd="%s/jdk/bin/keytool -noprompt -storepass changeit -import -trustcacerts -alias mongodb_operator_added_trust_ca_${RANDOM} -keystore %s/jdk/lib/security/cacerts" '/BEGIN/{close(cmd)};{print | cmd}' 2>&1 < %s/%s/ca-pem`, MMSHome, MMSHome, MMSHome, caCertVolume.Name)
 		postStart = func(lc *corev1.Lifecycle) {
 			if lc.PostStart == nil {
 				lc.PostStart = &corev1.Handler{Exec: &corev1.ExecAction{}}